fix(types): satisfy Next's route and layout contracts
These three errors only surface in .next/types (Next's generated validators),
so they are invisible to the CI typecheck, which does not build first, but a
local build turns them up. All three are real Next contract violations.

- api/speak-audio/route.ts exported six non-handler helpers (parseSpeakSource,
  isPrivateIP, parseRange, buildAudioResponse, isValidAuthor, isValidPermlink)
  so the unit test could reach them, but Next route modules may only export
  request handlers and known config. Moved them verbatim into a sibling
  _utils.ts (plus CodedError and ALLOWED_EXT that they share); route.ts and the
  spec import from there. No behaviour change.
- discover/layout.tsx declared a required communitiesDialog slot prop, but there
  is no @communitiesDialog parallel-route folder, so it was always undefined and
  rendering it was a no-op. Dropped the prop and the render.
- waves/layout.tsx declared a required view slot prop that no folder supplies and
  that the layout never rendered. Dropped it.

Verified against both states: tsc --noEmit is clean on a plain checkout and also
after pnpm build (when .next/types exists). Speak-audio route spec 52 passing.
